Better Implemented GridHelperScript's PaintInteractionRange function
This was a fun one, and one I'm going to continue to test rigorously. I made an algorithm to find all valid moveable tiles around a player's location taking into the player's movement range and the movement cost of the Tiles.
There is MORE going on here though.
In adding 'unpassable' Tiles, I have found Tiles that are unreachable being marked as Attackable and Walkable, when that isn't true. I'll have to see what I can do to validate this data further... Fun stuff! :D

:camel: