A Second, Closer InteractionRange Algorithm
Closer this time. I tried recursive approaches, one that I didn't save, and this one.
I will need to combine elements from the first, the recursion and the idea of a diminishing range (that implies a path / movement cost to the Tiles)  and the origin/manhattan distance calculations from this approach.
I'll have to figure out some way to do this. This is so fun, though!  :D
:camel: