Implementing basic Character + Unit Data, Items, Weapons, etc.
More needs to be done, and more thoroughly, to ensure I'm not missing anything, but I wanted to add these items to work more mechanics into my rough sketch before I start working on the game's flow.
With the whole Character thing, I need to work that out more in my mind to figure out what makes a Player's Character relevant vs an NPC besides their allegiance, which could just be a field or enumeration, as I've done it here.
So excited!
:camel: