ABOUT To Start on Selection Logic + A Character
This one's a big push, foolish me for doing so much.

First things first, I made a sprite to use as my character for now. Her name is Cynthia, and I literally made her up on the fly and animated a quick idle stance.

As well, I'm working on the Tile Selection logic that will be my FIRST foray into the
whole game flow thing. There's a lot of work to be done, but I'm starting with the Selection logic. I'll have to implement some functionality for soft selection when hovering over a Unit and all that stuff.

Overall, lots to figure out, but I'm still having an awesome time. How fun! There's a storm right now though so I'm trying not to get power surged.

:camel: