Refactoring Weapons->DamageItems +  Adding HealingItem SOs
I made some more data that organizes things that do damage, things that do healing, and things that do percentage-healing. This is all fun to make and it seems to be working out well, but I feel like there's better ways to architect this data.

In my mind I want there to be some distinction between a HealingItem being a potion and a HealingItem being some healing spell. I want to make the distinction but I want to avoid using inheritance JUST for the sake of avoiding duplication on the Range field for each of them. I might cave and do it, and refactor it out later if it's really that bad... I just want to make a distinction between something that's an Item and something that's a FieldItem with a range or something like that.
FieldItems, of course, being some item you can use on the field. This might make sorting out the Inventories easier? Hmm.

:camel: