:rotating_light: REFACTORING ITEMS WITH COMPOSITION /  INTERFACES :rotating_light:
my brain hurts... But this is for the better. Everywhere I go,  I hear about composition over inheritance, this and that, and the joy of using Interface-based design. However, I've come up with a solution that looks like it could benefit from this approach OVER inheritance with my Items and more. However, I want to use some of the interfaces to simulate abstraction to help me work around JUST using abstraction for types alone. An example is that a spell "NOSFERATU"  is currently a DamageItem with a DamageType, a range, etc, but this is technically the same as a bomb or a spear or something and isn't considered a spell except with different enums and this and that (which is how I've been doing things). If I were to make a few interfaces with properties that more accurately describe these items, I would be freaking SET! So I'm going to try to do this in the next commit. Just wanted to make note of this structural change.

:camel: