Refactored Items and BattleItems!
My brain MAY have hemorrhaged during this process BUT it was worth it. I remember that Inheritance can be dangerous, ESPECIALLY with data structures that may need to be changed here or there or whatever, and my use case for Inheritence in items wasn't really using it as it should be, only as a type.
Wanting more specificity than an enum but less of the ties of an inherited member, I went for some interfaces with Properties that define some common ground for what to expect with certain Items, BattleItems, etc. I also got to use interface inheritance to more sensibly implement interface-related duties :D

:camel: