A* Pathfinding Fixed!
There was a problem with my implementation before, in that there were certain tiles/GridNodes that would crash my program upon trying to find a path to, as well as openSet/closedSet sizes that seemed unbelievable, frankly, based on the amount of available squares.
So after a while I realized my error. I was using classes for GridNodes, so every time I used the List.Contains() method, I was comparing different instances of classes, NOT the data (inside of them which was duplicated horrendously). By overriding the Equals method in GridNode to check for equivalency, I now find that the pathfinding isn't giving me any errors, and is DRASTICALLY faster!
What a way to end the night! Perhaps I'll work more on control flow and UI stuff next time, now that I have a Queue<Vector3Int>-returning pathfinding function for my CharacterUnits to follow.
How fun!
:camel: