Implemented Passability in Pathfinding
So Cynthia accidentally walked through an impassable rock and I didn't like it so I fixed it!
As well, I implemented a movement speed float and used Time.deltaTime for CharacterUnitScript movement.

As I approach user flow and pulling up the main battle context menu after a character moves to a selectedTarget spot, I realize I may need to work with the Character Data with interfaces in the same way that I did with items to allow different permissions. ...Maybe, I'm still not sure if that's a good thing, using interfaces to help declare identity SOLELY using properties. But, it works well so far, and it plays nicely with the Unity inspector, so all's well currently.

:camel: