Updated Pathfinding Algorithm
Before, I could move directly to the positions of tiles that were ONLY  attackable. This was wrong, not to mention errors here and there that would be thrown if you wanted to do so on impassable terrain. I tried to fix these with some clarifications:
FindWalkablePathToTarget was a misleading name, since I got a path that wasn't always walkable. Instead, I need to find a path and ENSURE that it's all Walkable. If not, the path is no good, and we cannot find it.
As well, I don't need a Queue or Stack for my  path. Instead, I'll just take a List and validate it normally. So much for cool data structures  T-T.
Within the pathfinding algorithm, I've made it not care whether or not the start and endpoints are passable, ONLY about the neighbors and path there. If that's the case, we'll ALWAYS find and include the endpoint in the path, then cull any solely-Attackable tiles from the list at the end. How fun!

:camel: