Starting to Implement CSV item editing/importing.
There's a little custom editor script on the new 'UtilityGO' GameObject. There's also the new ItemDataBuilder, and in the Documents folder in Assets, the CSV and Excel files that will contain the Items. They're made to be somewhat intuitive, and I'm working on parsing the CSV file in a decent way. My mac is running low on battery though, so I want to commit this before it dies.

:camel: