Implementing a Singleton for my GMD210 class :D
I knew this'd show up here at some point, but I want to challenge myself to implement the patterns we learn in class somewhere in a project (this one, specifically). Thus, here I am, implementing an Audio system singleton. Though I want to do more with it (vertical layering / adaptive music tracks), this is sufficient for the due dates.

:camel: