Created TileSelectionStates and Manager
I have to go to class in a few, so I have to wrap up this current dev session here. Essentially, I started the state machine that will help take over the Tile Selection process that GridHelperScript formerly maintained for my proof-of-concept. Now, I'm figuring out how this new TileSelectionManager and its states will handle input and create data for a Move so we can put it in as a TurnActionCommand and all that jazz. There's still more to be done here, I'll likely keep working when I get home.

:camel: