Continuing to Implement CharacterSelectionState
11-4 TileSelectionState was my notes file for today.
So one thing I was able to do was get the initial PaintInteractionRange call and logic out of GridHelperScript. Again, in implementing this TileSelectionManager and its states, that's my primary goal for doing so. We'll return GridHelperScript to its original designation of helping with grid utilities and such with these TileSelectionStates. As well, I fixed an error with the PaintInteractionRange function stemming from the RecursivelyGetMoveRange function (literally one '!').
As I continue refactoring and reorganizing, I'll need to move files and reorganize my project more, which is what I've done with some scripts here.

:camel: