Implementing a lot of things (Up to ActionPromptState) [BIG update]
A LOT went down here, as they're my commits from my plane ride, and my car rides all without wifi where I couldn't upload smaller commits. Here's what I've done so far.

(My notes file was 'SNAKES ON THE PLANE' for the majority of these commits)

Majorly, I've been working and refining my approach with the TileSelection process, which has been going well.
As well, I've begun to implement arrow key and Return controls so far, though I've been meaning to create an InputManager script or something like that to help feed accepted inputs.
As well, I've achieved state reversion, and have taken notes on a bunch more steps of the TileSelection process that lead up to eventually performing TileActions.
I've implemented a better way to find things on tiles (OverlapBox-ing), and have finally begun to need an enemy unit to begin to test and debug with.

I'm really excited because this system is really at the heart of the game, being the main way you control it as the player. The AI will use this process as well (just without UI and maybe different AI-related states?) But as well, this is at the heart of a lot of systems in the game, eventually talking to the camera, audio system, inventory systems, BattleData systems, and more. It's key that I lay good foundation now.

:camel: