Fixed GetCharacterOnTile, cleaned up GridHelperScript somewhat
Back to smaller commits.
I found an error where I wasn't selected tiles by their position on the cell grid in the world, so I could select a tile by clicking on it and its upper, upper-right, and right neighbors. I fixed this with the help of some Gizmos, and now I'm back on track.

:camel: