Added more Music & SFX to TileSelectionProcess
The main purpose of this commit was to implement the Subclass Sandbox pattern for my GMD210 class. I did this by making some UI menu jingles and sounds for selection, confirmation, and declining the UI states. As you progress through the TileSelectionProcess, you will now hear little sounds that provide feedback to your selection that all of the TileSelectionState children share.

:camel: